third_party/blink/web_tests/platform/generic/accessibility/img-aria-button-alt-tag-expected.txt

This test makes sure that a img acting as an aria button still returns its alt tag as the description

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imgUIElement.name is 'alternate'
PASS successfullyParsed is true

TEST COMPLETE
